fix: Remove feature from modify interaction when deleting
Feature was present in selectedFeatures collection even after deleting it, which caused the mouse
pointer to stick to invisible features in Modify interaction. There was also a bug that we were
passing feature directly to event Subject instead of wrapping it in {feature, selector} structure
which was expected in subscriber. Modify interaction should be turned off when no feature is
selected too.

fix #1821