feat: Add a hack to style cluster layers using SLD
HACK is needed to style cluster layers. It wraps existing OL style function in a function which searches for for Text styles and in them for serialized feature arrays and instead sets the len
gth of this array as the label. If the geostyler text symbolizer had {{features}} as the text label template (which returns the "features" attribute of the parent/cluster feature) and returned
 '[object Object], [object Object]' the result would become "2".