feat(routing): Change routing provider to OpenRoutingService
Yournavigation.org was broken for some time already, so we needed to change it to some operational
one. OpenRoutingService provides 7000 free requests per day which makes it a good alternative for
small scale projects

BREAKING CHANGE: Be aware to specify OPENROUTESERVICE_API_KEY in .env of hslayers-server because
OpenRoutingService requests are routed through our proxy service to not expose the API key.

fix #1904