fix: Dont clear all GET params and rename hsl ones
Hslayers was incorrectly clearing all GET params in the browser url for applciation which includes
hslayers. We also want to keep track of the parameters which are used by hslayers but are not
regenerated continuously while panning the map (x, y, z) such as `composition` param. This commit
also provides a CONST list of all GET params which hsl uses.

fix #2098, #2084