refactor: Add panels dynamically extending panel base class
Now all panels in the panel space are loaded using the hsPanelContainerService, which originally
meant only for some custom extra panels. This allows adding only a subset of panels to the DOM and
also better treeshaling since not all Hslayers modules will be imported in the bundle.
HslayersModule still imports everything, but it will be possible to ommit importing of
HslayersModule in future.

ref #2111