fix: Remove basePath from url only in case of ngRouter
Previously basePath was removed from browser URL generated by hsl which was needed when ngRouter is
used in the container app, otherwise the base path would end up in url twice. This creates a bug
with wrong URL when base path is removed also when NgRouter is not used. I don't know how to detect
ngRouter in any other way than to specify its presence in HsConfig.

BREAKING CHANGE: Need to specify HsConfig.ngRouter = true if Angular router is used