feat: Cleanup after panels
When hslayers is reloaded in case of switching between routes the panels need to be reinitialized
and cleaned up. In case of a new/different panel-set it's not possible to clear ReplaySubject and
add new items, so we need to reinitialize the whole ReplaySubject object on destruction of panel
container component.

BREAKING CHANGE: Removal of hsLayoutService.initializedOnce variable which was a crutch anyway