feat: Map swipe state in url, properties stored
Now when sharing the map url, if map-swipe is available, the control state will be first checked from the url.

Also swipe control position and orientation is being saved to local storage.
So after the page reload, these values will be firstly checked from the storage.

SwipeSide layer property is also kept when sharing compositions

This commit also fixes query pop-up import path

Sets layer definition as WFS for vector layers only is it is saved to Layman

Adds a default style to a new layer if it is undefined