fix: Set copied layer props and source params before adding to map
This introduces origLayers layer property to keep original LAYERS when sub-layers
are used because there might be times when capabilities gets queried again and that
could reset origLayers to LAYERS, that only stores subLayers now thus breaking loading
of layers from cookies/composition json after page refresh.

In some cases we want the layers to be created, but not added to the map, like for copying.

addLayers function for OWS services gets refactored into several steps which can
be called independently.